requirements.txt
Django>=4.2
djangorestframework
djangorestframework-authtoken
Pillow
